+++ src/main/java/fll/web/report/AddAwardWinner.java
@@ -88,7 +88,7 @@ public class AddAwardWinner extends BaseFLLServlet {
           }
 
           winner = AwardWinners.getSubjectiveAwardWinner(connection, tournament.getTournamentID(), categoryTitle,
-                                                         teamNumber);
+                                                         awardGroup, teamNumber);
         } else if (EditAwardWinners.NON_NUMERIC_AWARD_TYPE.equals(awardType)) {
           final @Nullable NonNumericCategory category = challengeDescription.getNonNumericCategoryByTitle(categoryTitle);
           if (null == category) {
@@ -102,7 +102,7 @@ public class AddAwardWinner extends BaseFLLServlet {
               throw new FLLInternalException("Award group cannot be null for non-numeric award that is per award group");
             }
             winner = AwardWinners.getNonNumericAwardWinner(connection, tournament.getTournamentID(), categoryTitle,
-                                                           teamNumber);
+                                                           awardGroup, teamNumber);
           } else {
             winner = AwardWinners.getNonNumericOverallAwardWinner(connection, tournament.getTournamentID(),
                                                                   categoryTitle, teamNumber);
@@ -113,7 +113,7 @@ public class AddAwardWinner extends BaseFLLServlet {
           }
 
           winner = AwardWinners.getNonNumericAwardWinner(connection, tournament.getTournamentID(), categoryTitle,
-                                                         teamNumber);
+                                                         awardGroup, teamNumber);
         } else {
           throw new FLLInternalException("Unknown award type: '"
               + awardType
@@ -180,7 +180,7 @@ public class AddAwardWinner extends BaseFLLServlet {
           AwardWinners.updateSubjectiveAwardWinner(connection, tournament.getTournamentID(), winner);
         } else {
           if (null != AwardWinners.getSubjectiveAwardWinner(connection, tournament.getTournamentID(), categoryTitle,
-                                                            teamNumber)) {
+                                                            awardGroup, teamNumber)) {
             SessionAttributes.appendToMessage(session,
                                               String.format("<div class='error'>Team %d is already receiving award %s, cannot add a second time.</div>",
                                                             teamNumber, categoryTitle));
@@ -205,7 +205,7 @@ public class AddAwardWinner extends BaseFLLServlet {
             AwardWinners.updateNonNumericAwardWinner(connection, tournament.getTournamentID(), winner);
           } else {
             if (null != AwardWinners.getSubjectiveAwardWinner(connection, tournament.getTournamentID(), categoryTitle,
-                                                              teamNumber)) {
+                                                              awardGroup, teamNumber)) {
               SessionAttributes.appendToMessage(session,
                                                 String.format("<div class='error'>Team %d is already receiving award %s, cannot add a second time.</div>",
                                                               teamNumber, categoryTitle));
@@ -218,13 +218,13 @@ public class AddAwardWinner extends BaseFLLServlet {
           if (edit) {
             AwardWinners.updateNonNumericOverallAwardWinner(connection, tournament.getTournamentID(), winner);
           } else {
-            if (null != AwardWinners.getSubjectiveAwardWinner(connection, tournament.getTournamentID(), categoryTitle,
-                                                              teamNumber)) {
+            if (null != AwardWinners.getNonNumericOverallAwardWinner(connection, tournament.getTournamentID(),
+                                                                     categoryTitle, teamNumber)) {
               SessionAttributes.appendToMessage(session,
                                                 String.format("<div class='error'>Team %d is already receiving award %s, cannot add a second time.</div>",
                                                               teamNumber, categoryTitle));
             } else {
-              AwardWinners.addNonNumerciOverallAwardWinner(connection, tournament.getTournamentID(), winner);
+              AwardWinners.addNonNumericOverallAwardWinner(connection, tournament.getTournamentID(), winner);
             }
           }
         }
@@ -239,7 +239,7 @@ public class AddAwardWinner extends BaseFLLServlet {
           AwardWinners.updateNonNumericAwardWinner(connection, tournament.getTournamentID(), winner);
         } else {
           if (null != AwardWinners.getSubjectiveAwardWinner(connection, tournament.getTournamentID(), categoryTitle,
-                                                            teamNumber)) {
+                                                            awardGroup, teamNumber)) {
             SessionAttributes.appendToMessage(session,
                                               String.format("<div class='error'>Team %d is already receiving award %s, cannot add a second time.</div>",
                                                             teamNumber, categoryTitle));